Enable validation pipeline
For the validation pipeline to work we need the decorator metadata, so that
typescript's reflection feature can work properly. For this adding the
`babel-plugin-transform-typescript-metadata` plugin from Babel's
documentation[1]. Also this plugin didn't work well with typescript version 3,
hence bumping that as well. The issue on version 3 was that NestJS'
ValidationPipe with class-validator would raise error even if you provide all
the needed fields.

[1] - https://babeljs.io/docs/en/babel-plugin-transform-typescript#typescript-compiler-options